refactor: diviser NoteRoutes en UiRoutes et ApiRoutes pour respecter le SRP
