+++ notes-app/notes.json
@@ -1,10 +1,1 @@
-[
-  {
-    "id": "4fac18bc-f6a0-4abf-ba74-68849e1ce291",
-    "title": "Tagged",
-    "content": "Content",
-    "tags": [
-      "urgent"
-    ]
-  }
-]
+[]
